+++ History.md
@@ -1,5 +1,12 @@
 # Release History
 
+## 0.2 / 2011-10-10
+
+* Added ```store.opts.name``` to handle multiple stores by prefixing keys
+* Added ```store.removeAll()``` method
+* Fixed bug where only objects and arrays were persisted to webSQL and indexedDB
+* Updated documentation
+
 ## 0.1 / 2011-10-09
 
 * Initial release